requirements.txt
pandas
numpy
scikit-learn
Django
imblearn.over_sampling
tensorflow